+++ internal/echo/echo.go
@@ -43,9 +43,33 @@ func (m EchoRemote) ValidateParameters(parameters map[string]interface{}) error 
 }
 
 func (m EchoRemote) ListCommits(properties map[string]interface{}, parameters map[string]interface{}, tags []remote.Tag) ([]remote.Commit, error) {
-	panic("implement me")
+	res := []remote.Commit{{
+		Id:         "one",
+		Properties: map[string]interface{}{"tags": map[string]string{"name": "one"}, "timestamp": "2019-09-20T13:45:36Z"},
+	}, {
+		Id:         "two",
+		Properties: map[string]interface{}{"tags": map[string]string{"name": "two"}, "timestamp": "2019-09-20T13:45:37Z"},
+	}}
+	n := 0
+	for _, c := range res {
+		if remote.MatchTags(c.Properties, tags) {
+			res[n] = c
+			n++
+		}
+	}
+	res = res[:n]
+	remote.SortCommits(res)
+
+	return res, nil
 }
 
 func (m EchoRemote) GetCommit(properties map[string]interface{}, parameters map[string]interface{}, commitId string) (*remote.Commit, error) {
-	panic("implement me")
+	if commitId == "echo" {
+		return &remote.Commit{
+			Id:         "echo",
+			Properties: map[string]interface{}{"name": "echo", "timestamp": "2019-09-20T13:45:36Z"},
+		}, nil
+	} else {
+		return nil, nil
+	}
 }
